requirements.txt
Django==1.10.1
celery==3.1.23
redis==2.10.5
uWSGI==2.0.12
flake8==3.0.4
psycopg2==2.6.1
requests==2.11.1
djangorestframework==3.4.7
emoji==0.3.9